Move Header and Footer to Components folder
