+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import Navigation from "./nav";
-import Footer from "./footer";
+import Navigation from "../components/Header/Navigation";
+import Footer from "../components/Footer/Footer";
 import { Providers } from "@/components/provider/ProfileProvider";
 import { SpeedInsights } from "@vercel/speed-insights/next"
 import { Analytics } from '@vercel/analytics/react';
